LayoutTests/fast/dom/set-custom-validity-with-too-few-arguments.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test how setCustomValidity reacts to too few arguments.');

var b = document.createElement('button');
shouldThrow("b.setCustomValidity()");

var fs = document.createElement('fieldset');
shouldThrow("fs.setCustomValidity()");

var i = document.createElement('input');
shouldThrow("i.setCustomValidity()");

var k = document.createElement('keygen');
shouldThrow("k.setCustomValidity()");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
